separate build script into new file, apply cleanups
